resistance level set target added to virtualbike (not tested)

+++ src/bike.cpp
@@ -6,10 +6,10 @@ bike::bike()
 
 }
 
-void bike::changeResistance(double resistance) { requestResistance = resistance;}
+void bike::changeResistance(int8_t resistance) { requestResistance = resistance;}
 double bike::currentCrankRevolutions() { return CrankRevs;}
 uint16_t bike::lastCrankEventTime() { return LastCrankEventTime;}
-uint8_t bike::currentResistance() { return Resistance;}
+int8_t bike::currentResistance() { return Resistance;}
 uint8_t bike::currentCadence() { return Cadence;}
 uint8_t bike::fanSpeed() { return FanSpeed; }
 bool bike::connected() { return false; }
